add debug to multitps, update test_multitps_qt

+++ common/include/GMNR/common.h
@@ -7,7 +7,7 @@
 #include <Eigen/Sparse>
 
 namespace gmnr{
-	typedef float Scalar;
+	typedef double Scalar;
 	typedef Eigen::Matrix<Scalar, Eigen::Dynamic, Eigen::Dynamic> Matrix;
 	typedef Eigen::Matrix<Scalar, Eigen::Dynamic, 1> Vector;
